Refactoring code for better dev env #1
