+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.UrbanFarm"
         tools:targetApi="31">
         <activity
-            android:name=".activity.DetailTanamanPageActivity"
+            android:name=".activity.beranda.DetailTanamanPageActivity"
             android:exported="false" />
         <activity
             android:name=".activity.FragmentContainerActivity"
