+++ app/src/main/java/com/putragandad/urbanfarm/fragments/berandapage/BerandaPageFragment.kt
@@ -1,20 +1,16 @@
 package com.putragandad.urbanfarm.fragments.berandapage
 
-import android.content.Intent
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.GridLayoutManager
-import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.putragandad.urbanfarm.R
-import com.putragandad.urbanfarm.activity.DetailTanamanPageActivity
-import com.putragandad.urbanfarm.adapters.dashboard.ItemTanamanAdapter
-import com.putragandad.urbanfarm.databinding.ActivityFragmentContainerBinding
+import com.putragandad.urbanfarm.adapters.beranda.ItemTanamanAdapter
 import com.putragandad.urbanfarm.databinding.FragmentBerandaPageBinding
-import com.putragandad.urbanfarm.models.dashboard.ItemTanamanModels
+import com.putragandad.urbanfarm.models.beranda.ItemTanamanModels
 
 class BerandaPageFragment : Fragment() {
     private var _binding : FragmentBerandaPageBinding? = null
