+++ app/src/main/java/com/putragandad/urbanfarm/models/beranda/ItemTanamanModels.kt
@@ -1,4 +1,4 @@
-package com.putragandad.urbanfarm.models.dashboard
+package com.putragandad.urbanfarm.models.beranda
 
 data class ItemTanamanModels(val iconTanaman: Int,
                              val namaTanaman: String)
