+++ app/src/main/res/layout/activity_detail_tanaman_page.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.DetailTanamanPageActivity">
+    tools:context=".activity.beranda.DetailTanamanPageActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
